version 2 of the trajectory
